Create FindDuplicateGenotypes
This code returns sets of genotyped loci that are identical matches with other DNA vials. These are candidates for samples taken from the same individual as the chance of matching genotypes exactly at each locus for different individuals is extremely small (the probability of two individuals within the same grid and year having identical genotypes was found to be 1.17 x 10-19; J. Gorrell thesis).